Add CameraController script.
The camera will align itself to the centre of the two target
GameObjects (2 players) and will adjust it’s field of view to keep both
targets in view. The field of view will be clamped within a specified,
tweakable range. The fov adjustments are made using Lerp to ease in and
out.