Add PuzzleController script.
The PuzzleController keeps references to the two buttons and the
pillars involved in the puzzle. It then handles the events of either
button being pressed or released. This is where the logic for the
movement of the pillars is. The speed of pillar movement and the delays
can be set using the public vars in this script.